integraded new grbl redesign + status title updates every 50ms
