+++ Cargo.toml
@@ -12,3 +12,4 @@ toml = "0.5.8"
 serde = { version = "1.0.120", features = ["derive"] }
 iced = { version = "0.2", features = ["tokio"] }
 chrono = "0.4.19"
+regex = "1.4.3"
